refactor(vlog): merge root valueLog into engine/vlog.Consumer
Root-level vlog.go + vlog_gc.go were a tightly DB-coupled wrapper
sitting above engine/vlog.Manager. Move them into engine/vlog/ so the
top-level package keeps only the DB-side wiring (initVLog, head
bookkeeping, periodic GC driver).

- engine/vlog/consumer.go: Consumer (was valueLog), OpenConsumer,
  Read/Write/NewValuePtr, ReconcileManifest, ManagerFor/Managers,
  Logf, SetFileSize, Metrics, BackgroundCloser
- engine/vlog/discard.go: discardStats, EncodeDiscardStats,
  DecodeDiscardStats, flushDiscardStats, PopulateDiscardStats
- engine/vlog/gc.go: RunGC scheduler, FilterPendingDeletes,
  GCSampleSizeRatio/CountRatio, IteratorRefAdd/Del/Count
- ConsumerConfig (vlog tuning snapshot) + Deps (LSM interface +
  function-pointer callbacks for GetInternalEntry/BatchSet/
  SendToWriteCh/PeekShouldWriteValueToLSM) replace the old
  *DB-back-pointer; engine/vlog never imports root NoKV anymore
- HeadLogInterval / SmallCopyThreshold / DefaultDiscardStatsFlushThreshold
  / DiscardStatsKey exposed as exported package-level constants
- Eliminate the layering inversion where vlog.open mutated db.vheads:
  OpenConsumer now returns observed heads, db installs them
- Fix a subtle initVLog regression: lastLoggedHeads must be primed
  only from manifest heads (db.lsm.ValueLogHeadSnapshot()), never
  from the manager-initial heads observed during replay, otherwise
  the first updateHeadBuckets short-circuits shouldPersistHead and
  the bucket head never reaches the manifest
- Pure Consumer / discard / GC tests live alongside source in
  engine/vlog/{consumer,discard,gc}_test.go; the DB-coupled
  integration tests (Vlog/ValueLog/ValueGC/ManifestHead) consolidate
  into db_test.go alongside the shared opt/clearDir/newPointerEntry
  fixtures
