+++ db_test.go
@@ -19,8 +19,10 @@ import (
 	"github.com/feichai0017/NoKV/engine/lsm"
 	"github.com/feichai0017/NoKV/engine/manifest"
 	"github.com/feichai0017/NoKV/engine/vfs"
+	vlogpkg "github.com/feichai0017/NoKV/engine/vlog"
 	"github.com/feichai0017/NoKV/engine/wal"
 	dbruntime "github.com/feichai0017/NoKV/internal/runtime"
+	"github.com/feichai0017/NoKV/metrics"
 	myraft "github.com/feichai0017/NoKV/raft"
 	localmeta "github.com/feichai0017/NoKV/raftstore/localmeta"
 	raftmode "github.com/feichai0017/NoKV/raftstore/mode"
@@ -528,7 +530,7 @@ func TestGetValueLogEntryIsDetached(t *testing.T) {
 		value []byte
 	}{
 		{name: "small", value: bytes.Repeat([]byte("s"), 64)},
-		{name: "large", value: bytes.Repeat([]byte("l"), valueLogSmallCopyThreshold+512)},
+		{name: "large", value: bytes.Repeat([]byte("l"), vlogpkg.SmallCopyThreshold+512)},
 	} {
 		t.Run(tc.name, func(t *testing.T) {
 			opt := newTestOptions(t)
@@ -840,7 +842,7 @@ func TestRecoveryRemovesStaleValueLogSegment(t *testing.T) {
 		key := fmt.Appendf(nil, "key-%03d", i)
 		require.NoError(t, db.Set(key, val))
 	}
-	fids := db.vlog.managers[0].ListFIDs()
+	fids := db.vlog.Managers()[0].ListFIDs()
 	require.GreaterOrEqual(t, len(fids), 2)
 	staleFID := fids[0]
 
@@ -894,7 +896,7 @@ func TestRecoveryRemovesOrphanValueLogSegment(t *testing.T) {
 	val := make([]byte, 512)
 	require.NoError(t, db.Set(key, val))
 
-	headPtr := db.vlog.managers[0].Head()
+	headPtr := db.vlog.Managers()[0].Head()
 	require.False(t, headPtr.IsZero(), "expected value log head to be initialized")
 	headCopy := headPtr
 	require.NoError(t, db.lsm.LogValueLogHead(&headCopy))
@@ -919,7 +921,7 @@ func TestRecoveryRemovesOrphanValueLogSegment(t *testing.T) {
 	for id, meta := range status {
 		statusInfo[id] = meta.Valid
 	}
-	remainingFIDs := db2.vlog.managers[0].ListFIDs()
+	remainingFIDs := db2.vlog.Managers()[0].ListFIDs()
 
 	_, err := os.Stat(orphanPath)
 	require.Error(t, err)
@@ -1305,10 +1307,10 @@ func TestRecoverySkipsValueLogReplay(t *testing.T) {
 	userKey := []byte("vlog-replay-key")
 	internalKey := kv.InternalKey(kv.CFDefault, userKey, math.MaxUint64)
 	entry := kv.NewEntry(internalKey, []byte("payload"))
-	_, err := db.vlog.managers[0].AppendEntry(entry)
+	_, err := db.vlog.Managers()[0].AppendEntry(entry)
 	require.NoError(t, err)
 	entry.DecrRef()
-	require.NoError(t, db.vlog.managers[0].SyncActive())
+	require.NoError(t, db.vlog.Managers()[0].SyncActive())
 	require.NoError(t, db.Close())
 
 	db2 := openTestDB(t, opt)
@@ -1605,7 +1607,7 @@ func drForEachMemTableEngine(t *testing.T, fn func(t *testing.T, engine MemTable
 func drSimulateCrash(t *testing.T, db *DB) {
 	t.Helper()
 	_ = db.Info().close()
-	for _, mgr := range db.vlog.managers {
+	for _, mgr := range db.vlog.Managers() {
 		if mgr != nil {
 			_ = mgr.Close()
 		}
@@ -2717,4 +2719,518 @@ func TestDBValueLogEnabledRoundTrip(t *testing.T) {
 
 	_, err = os.Stat(filepath.Join(dir, "vlog"))
 	require.NoError(t, err, "vlog dir must exist when EnableValueLog is true")
+}
+
+// opt is the shared test-fixture Options used by the legacy vlog tests
+// and by stats_test.go / iterator_test.go / db_test.go fast-path tests.
+// Tests that mutate it must restore the previous value in a defer.
+var opt = &Options{
+	WorkDir:             "./work_test",
+	SSTableMaxSz:        1 << 10,
+	MemTableSize:        1 << 10,
+	EnableValueLog:      true,
+	ValueLogFileSize:    1 << 20,
+	ValueThreshold:      0,
+	ValueLogBucketCount: 1,
+	MaxBatchCount:       10,
+	MaxBatchSize:        1 << 20,
+	ThermosEnabled:      true,
+	ThermosBits:         8,
+	ThermosTopK:         8,
+}
+
+// clearDir wipes the shared opt.WorkDir between tests and re-points it
+// at a fresh temp directory. Used by stats_test.go and iterator_test.go
+// in addition to the vlog tests below.
+func clearDir() {
+	if opt == nil {
+		return
+	}
+	if opt.WorkDir != "" {
+		_ = os.RemoveAll(opt.WorkDir)
+	}
+	dir, err := os.MkdirTemp("", "nokv-vlog-test-")
+	if err != nil {
+		panic(err)
+	}
+	opt.WorkDir = dir
+}
+
+func newPointerEntry(userKey, value []byte) *kv.Entry {
+	return kv.NewInternalEntry(kv.CFDefault, userKey, nonTxnMaxVersion, value, kv.BitValuePointer, 0)
+}
+
+func keyForBucket(t *testing.T, bucket int, buckets int) []byte {
+	t.Helper()
+	for i := range 10000 {
+		userKey := fmt.Appendf(nil, "gc-bucket-key-%d", i)
+		internal := kv.InternalKey(kv.CFDefault, userKey, 1)
+		if kv.ValueLogBucket(internal, uint32(buckets)) == uint32(bucket) {
+			return userKey
+		}
+	}
+	t.Fatalf("unable to find key for bucket %d", bucket)
+	return nil
+}
+
+func newRandEntry(sz int) *kv.Entry {
+	v := make([]byte, sz)
+	rng := rand.New(rand.NewSource(time.Now().UnixNano()))
+	_, _ = rng.Read(v[:rng.Intn(sz)])
+	key := fmt.Appendf(nil, "vlog-rand-%d-%d", time.Now().UnixNano(), rng.Uint64())
+	e := kv.NewEntry(key, v)
+	e.ExpiresAt = uint64(time.Now().Add(12 * time.Hour).Unix())
+	return e
+}
+
+func getItemValue(t *testing.T, item *kv.Entry) (val []byte) {
+	t.Helper()
+	if item == nil {
+		return nil
+	}
+	var v []byte
+	v = append(v, item.Value...)
+	if v == nil {
+		return nil
+	}
+	return v
+}
+
+func TestVlogBase(t *testing.T) {
+	clearDir()
+	db := openTestDB(t, opt)
+	defer func() { _ = db.Close() }()
+	log := db.vlog
+	const val1 = "sampleval012345678901234567890123"
+	const val2 = "samplevalb012345678901234567890123"
+	require.True(t, int64(len(val1)) >= db.opt.ValueThreshold)
+
+	e1 := newPointerEntry([]byte("samplekey"), []byte(val1))
+	e2 := newPointerEntry([]byte("samplekeyb"), []byte(val2))
+
+	b := new(dbruntime.Request)
+	b.Entries = []*kv.Entry{e1, e2}
+
+	require.NoError(t, log.Write([]*dbruntime.Request{b}))
+	e1.DecrRef()
+	e2.DecrRef()
+	require.Len(t, b.Ptrs, 2)
+	t.Logf("Pointer written: %+v %+v\n", b.Ptrs[0], b.Ptrs[1])
+
+	mgr1, err := log.ManagerFor(b.Ptrs[0].Bucket)
+	require.NoError(t, err)
+	mgr2, err := log.ManagerFor(b.Ptrs[1].Bucket)
+	require.NoError(t, err)
+	payload1, unlock1, err1 := mgr1.Read(&b.Ptrs[0])
+	payload2, unlock2, err2 := mgr2.Read(&b.Ptrs[1])
+	require.NoError(t, err1)
+	require.NoError(t, err2)
+	if unlock1 != nil {
+		defer unlock1()
+	}
+	if unlock2 != nil {
+		defer unlock2()
+	}
+	entry1, err := kv.DecodeEntry(payload1)
+	require.NoError(t, err)
+	defer entry1.DecrRef()
+	entry2, err := kv.DecodeEntry(payload2)
+	require.NoError(t, err)
+	defer entry2.DecrRef()
+
+	_, key1, ts1, ok := kv.SplitInternalKey(entry1.Key)
+	require.True(t, ok)
+	require.Equal(t, []byte("samplekey"), key1)
+	require.Equal(t, nonTxnMaxVersion, ts1)
+	require.Equal(t, []byte(val1), entry1.Value)
+	require.Equal(t, kv.BitValuePointer, entry1.Meta)
+
+	_, key2, ts2, ok := kv.SplitInternalKey(entry2.Key)
+	require.True(t, ok)
+	require.Equal(t, []byte("samplekeyb"), key2)
+	require.Equal(t, nonTxnMaxVersion, ts2)
+	require.Equal(t, []byte(val2), entry2.Value)
+	require.Equal(t, kv.BitValuePointer, entry2.Meta)
+}
+
+func TestVersionedEntryValueLogPointer(t *testing.T) {
+	clearDir()
+	prevThreshold := opt.ValueThreshold
+	prevFileSize := opt.ValueLogFileSize
+	opt.ValueThreshold = 0
+	opt.ValueLogFileSize = 1 << 20
+	defer func() {
+		opt.ValueThreshold = prevThreshold
+		opt.ValueLogFileSize = prevFileSize
+	}()
+
+	db := openTestDB(t, opt)
+	defer func() { _ = db.Close() }()
+
+	key := []byte("versioned-vlog")
+	version := uint64(7)
+	value := bytes.Repeat([]byte("v"), 64)
+
+	applyVersionedEntryForTest(t, db, kv.CFDefault, key, version, value, 0)
+	entry, err := db.GetInternalEntry(kv.CFDefault, key, version)
+	require.NoError(t, err)
+	require.Equal(t, kv.CFDefault, entry.CF)
+	_, userKey, _, ok := kv.SplitInternalKey(entry.Key)
+	require.True(t, ok)
+	require.Equal(t, key, userKey)
+	require.Equal(t, version, kv.Timestamp(entry.Key))
+	require.Equal(t, value, entry.Value)
+	entry.DecrRef()
+}
+
+func TestVlogSyncWritesCoversAllSegments(t *testing.T) {
+	clearDir()
+
+	prevSync := opt.SyncWrites
+	prevThreshold := opt.ValueThreshold
+	prevFileSize := opt.ValueLogFileSize
+	opt.SyncWrites = true
+	opt.ValueThreshold = 0
+	opt.ValueLogFileSize = 256
+	defer func() {
+		opt.SyncWrites = prevSync
+		opt.ValueThreshold = prevThreshold
+		opt.ValueLogFileSize = prevFileSize
+	}()
+
+	db := openTestDB(t, opt)
+	defer func() { _ = db.Close() }()
+	log := db.vlog
+
+	payload := bytes.Repeat([]byte("v"), 180)
+	e1 := newPointerEntry([]byte("sync-key-1"), payload)
+	e2 := newPointerEntry([]byte("sync-key-2"), payload)
+	req := &dbruntime.Request{Entries: []*kv.Entry{e1, e2}}
+
+	require.NoError(t, log.Write([]*dbruntime.Request{req}))
+	e1.DecrRef()
+	e2.DecrRef()
+
+	if len(req.Ptrs) != 2 {
+		t.Fatalf("expected 2 value pointers, got %d", len(req.Ptrs))
+	}
+	if req.Ptrs[0].Fid == req.Ptrs[1].Fid && req.Ptrs[0].Bucket == req.Ptrs[1].Bucket {
+		t.Fatalf("expected pointers in different vlog segments or buckets, got fid=%d bucket=%d", req.Ptrs[0].Fid, req.Ptrs[0].Bucket)
+	}
+}
+
+func TestValueGC(t *testing.T) {
+	clearDir()
+	opt.ValueLogFileSize = 1 << 20
+	origCompactors := opt.NumCompactors
+	origMemTableSize := opt.MemTableSize
+	origSSTableMaxSz := opt.SSTableMaxSz
+	opt.NumCompactors = 0
+	opt.MemTableSize = 8 << 20
+	opt.SSTableMaxSz = 8 << 20
+	db := openTestDB(t, opt)
+	defer func() { _ = db.Close() }()
+	defer func() {
+		opt.NumCompactors = origCompactors
+		opt.MemTableSize = origMemTableSize
+		opt.SSTableMaxSz = origSSTableMaxSz
+	}()
+	sz := 32 << 10
+	kvList := make([]*kv.Entry, 0, 100)
+	defer func() {
+		for _, e := range kvList {
+			e.DecrRef()
+		}
+	}()
+
+	for range 100 {
+		e := newRandEntry(sz)
+		eCopy := kv.NewEntry(e.Key, e.Value)
+		eCopy.Meta = e.Meta
+		eCopy.ExpiresAt = e.ExpiresAt
+		kvList = append(kvList, eCopy)
+
+		if e.ExpiresAt > 0 {
+			ttl := time.Until(time.Unix(int64(e.ExpiresAt), 0))
+			require.NoError(t, db.SetWithTTL(e.Key, e.Value, ttl))
+		} else {
+			require.NoError(t, db.Set(e.Key, e.Value))
+		}
+		e.DecrRef()
+	}
+	if err := db.RunValueLogGC(0.9); err != nil && !errors.Is(err, utils.ErrNoRewrite) {
+		require.NoError(t, err)
+	}
+	for _, e := range kvList {
+		item, err := db.Get(e.Key)
+		require.NoErrorf(t, err, "missing key after gc user_key=%q raw=%x", string(e.Key), e.Key)
+		val := getItemValue(t, item)
+		require.NotNil(t, val)
+		require.True(t, bytes.Equal(item.Key, e.Key), "key not equal: e:%s, v:%s", e.Key, item.Key)
+		require.True(t, bytes.Equal(item.Value, e.Value), "value not equal: e:%s, v:%s", e.Value, item.Key)
+	}
+}
+
+func TestValueLogGCParallelScheduling(t *testing.T) {
+	cfg := *opt
+	cfg.WorkDir = t.TempDir()
+	cfg.ValueThreshold = 0
+	cfg.ValueLogBucketCount = 2
+	cfg.ValueLogFileSize = 4 << 10
+	cfg.ValueLogGCParallelism = 2
+	cfg.ValueLogGCInterval = 0
+	cfg.NumCompactors = 2
+
+	db := openTestDB(t, &cfg)
+	defer func() { _ = db.Close() }()
+
+	key0 := keyForBucket(t, 0, cfg.ValueLogBucketCount)
+	key1 := keyForBucket(t, 1, cfg.ValueLogBucketCount)
+	payload := bytes.Repeat([]byte("x"), 512)
+
+	hasSealed := func() bool {
+		for bucket := 0; bucket < cfg.ValueLogBucketCount; bucket++ {
+			mgr, err := db.vlog.ManagerFor(uint32(bucket))
+			require.NoError(t, err)
+			if len(mgr.ListFIDs()) < 2 {
+				return false
+			}
+		}
+		return true
+	}
+
+	for i := 0; i < 200 && !hasSealed(); i++ {
+		require.NoError(t, db.Set(key0, payload))
+		require.NoError(t, db.Set(key1, payload))
+	}
+	require.True(t, hasSealed(), "expected sealed vlog segments in each bucket")
+
+	metrics.ResetValueLogGCMetricsForTesting()
+	before := metrics.DefaultValueLogGCCollector().Snapshot().GCScheduled
+
+	_ = db.RunValueLogGC(0.99)
+
+	after := metrics.DefaultValueLogGCCollector().Snapshot().GCScheduled
+	if after-before < 2 {
+		t.Fatalf("expected parallel GC scheduling, delta=%d", after-before)
+	}
+}
+
+func TestValueLogIterateReleasesEntries(t *testing.T) {
+	clearDir()
+	db := openTestDB(t, opt)
+	defer func() { _ = db.Close() }()
+
+	val := bytes.Repeat([]byte("x"), 128)
+	entry := kv.NewEntry([]byte("iter-key"), val)
+	require.NoError(t, db.Set(entry.Key, entry.Value))
+	entry.DecrRef()
+
+	managers := db.vlog.Managers()
+	active := managers[0].ActiveFID()
+
+	var captured []*kv.Entry
+	_, err := managers[0].Iterate(active, kv.ValueLogHeaderSize, func(e *kv.Entry, vp *kv.ValuePtr) error {
+		captured = append(captured, e)
+		return nil
+	})
+	require.NoError(t, err)
+	require.NotZero(t, len(captured), "expected to capture at least one entry")
+
+	for _, e := range captured {
+		if len(e.Key) != 0 || len(e.Value) != 0 {
+			t.Fatalf("expected entry to be reset after DecrRef")
+		}
+	}
+}
+
+func TestDecodeWalEntryReleasesEntries(t *testing.T) {
+	orig := kv.NewEntry([]byte("decode-key"), []byte("decode-val"))
+	buf := &bytes.Buffer{}
+	payload, err := kv.EncodeEntry(buf, orig)
+	require.NoError(t, err)
+	orig.DecrRef()
+
+	entry, err := kv.DecodeEntry(payload)
+	require.NoError(t, err)
+	entry.DecrRef()
+
+	if len(entry.Key) != 0 || len(entry.Value) != 0 {
+		t.Fatalf("expected decoded entry to reset after DecrRef")
+	}
+}
+
+func TestValueLogWriteAppendFailureRewinds(t *testing.T) {
+	clearDir()
+	cfg := *opt
+	cfg.ValueLogFileSize = 256
+	injected := errors.New("append failure")
+	rotatePath := filepath.Join(cfg.WorkDir, "vlog", "bucket-000", "00001.vlog")
+	cfg.FS = vfs.NewFaultFSWithPolicy(vfs.OSFS{}, vfs.NewFaultPolicy(
+		vfs.FailOnceRule(vfs.OpOpenFile, rotatePath, injected),
+	))
+	db := openTestDB(t, &cfg)
+	defer func() { _ = db.Close() }()
+
+	mgr, err := db.vlog.ManagerFor(0)
+	require.NoError(t, err)
+	head := mgr.Head()
+
+	req := dbruntime.RequestPool.Get().(*dbruntime.Request)
+	req.Reset()
+	entries := []*kv.Entry{
+		newPointerEntry([]byte("afail"), bytes.Repeat([]byte("a"), 512)),
+		newPointerEntry([]byte("bfail"), bytes.Repeat([]byte("b"), 512)),
+	}
+	req.LoadEntries(entries)
+	req.IncrRef()
+	defer req.DecrRef()
+
+	err = db.vlog.Write([]*dbruntime.Request{req})
+	require.Error(t, err)
+	require.ErrorIs(t, err, injected)
+	require.Equal(t, head, mgr.Head())
+	require.Len(t, req.Ptrs, 0)
+}
+
+func TestValueLogWriteRotateFailureRewinds(t *testing.T) {
+	clearDir()
+	cfg := *opt
+	cfg.ValueLogFileSize = 256
+	injected := errors.New("rotate failure")
+	rotatePath := filepath.Join(cfg.WorkDir, "vlog", "bucket-000", "00001.vlog")
+	cfg.FS = vfs.NewFaultFSWithPolicy(vfs.OSFS{}, vfs.NewFaultPolicy(
+		vfs.FailOnceRule(vfs.OpOpenFile, rotatePath, injected),
+	))
+	db := openTestDB(t, &cfg)
+	defer func() { _ = db.Close() }()
+
+	mgr, err := db.vlog.ManagerFor(0)
+	require.NoError(t, err)
+	head := mgr.Head()
+
+	req := dbruntime.RequestPool.Get().(*dbruntime.Request)
+	req.Reset()
+	entries := []*kv.Entry{
+		newPointerEntry([]byte("rfail1"), bytes.Repeat([]byte("x"), 512)),
+		newPointerEntry([]byte("rfail2"), bytes.Repeat([]byte("y"), 512)),
+	}
+	req.LoadEntries(entries)
+	req.IncrRef()
+	defer req.DecrRef()
+
+	err = db.vlog.Write([]*dbruntime.Request{req})
+	require.Error(t, err)
+	require.ErrorIs(t, err, injected)
+	require.Equal(t, head, mgr.Head())
+	require.Len(t, req.Ptrs, 0)
+}
+
+func TestValueLogWriteInlineRequestSkipsPtrs(t *testing.T) {
+	clearDir()
+	prevThreshold := opt.ValueThreshold
+	opt.ValueThreshold = 1 << 20
+	defer func() { opt.ValueThreshold = prevThreshold }()
+
+	db := openTestDB(t, opt)
+	defer func() { _ = db.Close() }()
+
+	req := dbruntime.RequestPool.Get().(*dbruntime.Request)
+	req.Reset()
+	entry := kv.NewInternalEntry(kv.CFDefault, []byte("inline-vlog"), nonTxnMaxVersion, []byte("v"), 0, 0)
+	req.LoadEntries([]*kv.Entry{entry})
+	req.IncrRef()
+	defer req.DecrRef()
+
+	require.NoError(t, db.vlog.Write([]*dbruntime.Request{req}))
+	require.Len(t, req.Ptrs, 0)
+	require.Len(t, req.PtrIdxs, 0)
+	require.Len(t, req.PtrBuckets, 0)
+}
+
+func TestValueLogReadCopiesSmallValue(t *testing.T) {
+	clearDir()
+	prevThreshold := opt.ValueThreshold
+	opt.ValueThreshold = 0
+	defer func() { opt.ValueThreshold = prevThreshold }()
+
+	db := openTestDB(t, opt)
+	defer func() { _ = db.Close() }()
+
+	entry := kv.NewInternalEntry(kv.CFDefault, []byte("small-read"), nonTxnMaxVersion, []byte("v"), 0, 0)
+	vp, err := db.vlog.NewValuePtr(entry)
+	entry.DecrRef()
+	require.NoError(t, err)
+
+	val, cb, err := db.vlog.Read(vp)
+	require.NoError(t, err)
+	require.Nil(t, cb)
+	require.Equal(t, []byte("v"), val)
+}
+
+func TestManifestHeadMatchesValueLogHead(t *testing.T) {
+	clearDir()
+	opt.ValueThreshold = 0
+	db := openTestDB(t, opt)
+	defer func() { _ = db.Close() }()
+
+	entry := kv.NewEntry([]byte("manifest-head"), []byte("value"))
+	entry.Key = kv.InternalKey(kv.CFDefault, entry.Key, math.MaxUint32)
+	if err := db.batchSet([]*kv.Entry{entry}); err != nil {
+		t.Fatalf("batchSet: %v", err)
+	}
+
+	mgr, err := db.vlog.ManagerFor(0)
+	require.NoError(t, err)
+	head := mgr.Head()
+	heads := db.getHeads()
+	meta, ok := heads[0]
+	if !ok {
+		t.Fatalf("expected manifest head")
+	}
+	if meta.Fid != head.Fid {
+		t.Fatalf("manifest fid %d does not match manager %d", meta.Fid, head.Fid)
+	}
+	if meta.Offset != head.Offset {
+		t.Fatalf("manifest offset %d does not match manager %d", meta.Offset, head.Offset)
+	}
+}
+
+func TestValueLogGCSkipBlocked(t *testing.T) {
+	clearDir()
+	opt := NewDefaultOptions()
+	opt.ValueLogFileSize = 1 << 20
+	opt.NumCompactors = 0
+	db := openTestDB(t, opt)
+	defer func() { _ = db.Close() }()
+
+	e := kv.NewEntry([]byte("gc-skip"), []byte("v"))
+	require.NoError(t, db.Set(e.Key, e.Value))
+	e.DecrRef()
+
+	db.applyThrottle(lsm.WriteThrottleStop)
+	defer db.applyThrottle(lsm.WriteThrottleNone)
+
+	if err := db.RunValueLogGC(0.5); err != nil && !errors.Is(err, utils.ErrNoRewrite) {
+		t.Fatalf("expected ErrNoRewrite when writes blocked, got %v", err)
+	}
+}
+
+func TestValueLogPopulateDiscardStatsLoadsPersistedEntry(t *testing.T) {
+	opt := newTestOptions(t)
+	db := openTestDB(t, opt)
+	defer func() { _ = db.Close() }()
+
+	stats := map[manifest.ValueLogID]int64{
+		{Bucket: 0, FileID: 7}: 99,
+		{Bucket: 1, FileID: 4}: 12,
+	}
+	encoded, err := vlogpkg.EncodeDiscardStats(stats)
+	require.NoError(t, err)
+
+	entry := kv.NewInternalEntry(kv.CFDefault, vlogpkg.DiscardStatsKey, nonTxnMaxVersion, encoded, 0, 0)
+	defer entry.DecrRef()
+	require.NoError(t, db.ApplyInternalEntries([]*kv.Entry{entry}))
+
+	require.NoError(t, db.vlog.PopulateDiscardStats())
 }
